refactor: Studio V3 Phase 1-4 — foundation, hooks, context, component decomposition
Complete Phase 1-4 of the Studio V3 refactor plan:

Phase 1 (Foundation):
- Add getCapabilityFieldType() mapping for CapabilityForm
- Add GenerationConfigSchema (unified image+video Zod schema)
- Tighten quality field to z.enum(['auto','low','medium','high'])

Phase 4 (Component Decomposition):
- Extract 6 sub-components from StudioWorkspace (615→75 lines):
  StudioCardSelectors, StudioPreview, StudioToolbarPanels,
  StudioCardManagement, StudioProjectHistory, StudioVideoMode
- Each sub-component consumes split contexts directly (zero prop drilling)
- Update plan with implementation progress tracking

Phases 2-3 were already implemented in prior commits.
