+++ src/components/business/studio/index.ts
@@ -2,3 +2,9 @@ export { StudioModeSelector } from './StudioModeSelector'
 export { StudioPromptArea } from './StudioPromptArea'
 export { StudioGenerateBar } from './StudioGenerateBar'
 export { StudioErrorBoundary } from './StudioErrorBoundary'
+export { StudioCardSelectors } from './StudioCardSelectors'
+export { StudioPreview } from './StudioPreview'
+export { StudioToolbarPanels } from './StudioToolbarPanels'
+export { StudioCardManagement } from './StudioCardManagement'
+export { StudioProjectHistory } from './StudioProjectHistory'
+export { StudioVideoMode } from './StudioVideoMode'
